Added further existence tests of the pieces on the board.
Added const to the existence functions. Have the simple moves created,
just need to work on the jump now. Also, need to check for bounds on the
move functions so that they do not reference memory that is not apart of
the board. Maybe need to add a buffer around to the board for this.