Modified checker_board.hpp and game_piece.hpp.
Fixed the memory bug, it was a missing return statement in the
Game_Piece class function get_moves.

Decided that the map should be switched over to coordinates to help with
not having to translate from pairs to strings and vice versa. Thus, a
number of functions just needed a parameter change.