Restructured testing of the program.
Separated the Catch main definition into its own file in order to speed
up the compilation of the testing suite.

	There are now separate files for each aspects that will need
	testing. Currently there is only 2, one for the neural network
	class and one for the network specification class.

There is currently an issue with the Makefile in that the testing suite
is having trouble with multiple includes. Also, need to figure out how
to compile the implementation files along with the test files.