Update to the node class.
Move the sigmoid function into the private section of the class.

Renamed member variable input to value. Because you don't need to track
the value that was passed to the node, just what the summation of input
multiplied by the node's weight.

Serialization of the class only saves the weights of the node now.